+++ packages/@expo/cli/src/start/server/metro/__tests__/withMetroMultiPlatform.test.ts
@@ -163,6 +163,110 @@ describe(withExtendedResolver, () => {
     );
   });
 
+  it(`resolves to @expo/vector-icons on any platform`, async () => {
+    vol.fromJSON(
+      {
+        'node_modules/@react-native/assets-registry/registry.js': '',
+        'node_modules/@expo/vector-icons/index.js': '',
+      },
+      '/'
+    );
+
+    ['ios', 'web'].forEach((platform) => {
+      const modified = withExtendedResolver(asMetroConfig({ projectRoot: '/' }), {
+        platforms: ['ios', 'web'],
+        tsconfig: { baseUrl: '/src', paths: { '/*': ['*'] } },
+        isTsconfigPathsEnabled: true,
+      });
+
+      modified.resolver.resolveRequest!(
+        getDefaultRequestContext(),
+        'react-native-vector-icons',
+        platform
+      );
+
+      expect(getResolveFunc()).toBeCalledWith(expect.anything(), '@expo/vector-icons', platform);
+    });
+  });
+
+  it(`resolves nested imports to @expo/vector-icons on any platform`, async () => {
+    vol.fromJSON(
+      {
+        'node_modules/@react-native/assets-registry/registry.js': '',
+        'node_modules/@expo/vector-icons/index.js': '',
+      },
+      '/'
+    );
+
+    ['ios', 'web'].forEach((platform) => {
+      const modified = withExtendedResolver(asMetroConfig({ projectRoot: '/' }), {
+        platforms: ['ios', 'web'],
+        tsconfig: { baseUrl: '/src', paths: { '/*': ['*'] } },
+        isTsconfigPathsEnabled: true,
+      });
+
+      modified.resolver.resolveRequest!(
+        getDefaultRequestContext(),
+        'react-native-vector-icons/FontAwesome',
+        platform
+      );
+
+      expect(getResolveFunc()).toBeCalledWith(
+        expect.anything(),
+        '@expo/vector-icons/FontAwesome',
+        platform
+      );
+    });
+  });
+
+  it(`does not alias react-native-vector-icons if @expo/vector-icons is not installed`, async () => {
+    vol.fromJSON(
+      {
+        'node_modules/@react-native/assets-registry/registry.js': '',
+      },
+      '/'
+    );
+
+    ['ios', 'web'].forEach((platform) => {
+      const modified = withExtendedResolver(asMetroConfig({ projectRoot: '/' }), {
+        platforms: ['ios', 'web'],
+        tsconfig: { baseUrl: '/src', paths: { '/*': ['*'] } },
+        isTsconfigPathsEnabled: true,
+      });
+
+      modified.resolver.resolveRequest!(
+        getDefaultRequestContext(),
+        'react-native-vector-icons',
+        platform
+      );
+
+      expect(getResolveFunc()).toBeCalledWith(
+        expect.anything(),
+        'react-native-vector-icons',
+        platform
+      );
+    });
+  });
+
+  it(`allows importing @expo/vector-icons`, async () => {
+    vol.fromJSON(
+      {
+        'node_modules/@react-native/assets-registry/registry.js': '',
+        'node_modules/@expo/vector-icons/index.js': '',
+      },
+      '/'
+    );
+    const platform = 'ios';
+    const modified = withExtendedResolver(asMetroConfig({ projectRoot: '/' }), {
+      platforms: ['ios', 'web'],
+      tsconfig: { baseUrl: '/src', paths: { '/*': ['*'] } },
+      isTsconfigPathsEnabled: true,
+    });
+
+    modified.resolver.resolveRequest!(getDefaultRequestContext(), '@expo/vector-icons', platform);
+    expect(getResolveFunc()).toBeCalledWith(expect.anything(), '@expo/vector-icons', platform);
+  });
+
   it(`resolves a node.js built-in as a shim on web`, async () => {
     mockMinFs();